feat: Rename `Block` to `ProvenBlock`
